refactor: rename classes to match diagrams
